SPCreporter: Creates Metric Reports using Statistical Process Control in the NHS style

Takes a dataset file and a configuration file to produce an HTML report with each metric presented as an SPC chart. The initial rendering of the report contains the charts folded to enable quick scanning of "variation" and "assurance" status. Each metric can be unfolded to reveal the SPC chart to support clinical and management planning. The NHSRplotthedots package is used for SPC calculations, and the "variation" and "assurance" is in line with the popular standard within the UK's National Health Service (NHS).

See also

Author

Maintainer: Tom Smith tomsmith_uk@hotmail.com

Authors:

Fran Barton fbarton@alwaysdata.net